+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Allow `license-file` to be a list
 
 ## [0.17.0] - 2017-01-30
 ### Added